Make sure a file exists before deleting it, and even then wrap the delete
2007-12-28  Gabriel Burt  <gabriel.burt@gmail.com>

	* src/Core/Banshee/PlayerInterface.cs: Make sure a file exists before
	deleting it, and even then wrap the delete in a try/catch.  Patch from
	Wade Menard, fixes BGO #484009.

svn path=/branches/banshee/stable/; revision=2929

+++ src/Core/Banshee/PlayerInterface.cs
@@ -1924,7 +1924,11 @@ namespace Banshee
 
         private void DeleteSong(TrackInfo ti)
         {
-            File.Delete(ti.Uri.LocalPath);
+            if (File.Exists(ti.Uri.LocalPath)) {
+                try {
+                    File.Delete(ti.Uri.LocalPath);
+                } catch(Exception) { }
+            }
 
             // trim empty parent directories
             try {
